# api/stocks_funds_purchases.go
// Copyright (c) 2020, Marcelo Jorge Vieira (https://github.com/mfinancecombr)
// Licensed under the BSD 3-Clause License

package api

import (
	"net/http"

	"github.com/labstack/echo/v4"
	"github.com/mfinancecombr/finance-wallet-api/wallet"
	log "github.com/sirupsen/logrus"
)

func (s *server) getAllStockFundsPurchases(c echo.Context) error {
	log.Debug("[API] Retrieving all stocks funds purchases")
	result, err := s.db.GetAllStocksFundsPurchases()
	if err != nil {
		log.Errorf("[API] Error on retrieve data: %v", err)
		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
	}
	return c.JSON(http.StatusOK, result)
}

// getStockFundPurchaseByID godoc
// @Summary Get stocks fund purchase by ID
// @Description get stocks fund purchase data
// @Accept json
// @Produce json
// @Success 200 {object} wallet.StockFund
// @Router /stocks-funds/purchases/{id} [get]
// @Param id path string true "Purchase id"
func (s *server) getStockFundPurchaseByID(c echo.Context) error {
	id := c.Param("id")
	log.Debugf("[API] Retrieving stock fund purchase with id: %s", id)
	result, err := s.db.GetStockFundPurchaseByID(id)
	if err != nil {
		log.Errorf("[API] Error on retrieve data: %v", err)
		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
	}
	if result == nil {
		return c.JSON(http.StatusNotFound, "Stock fund purchase data not found")
	}
	return c.JSON(http.StatusOK, result)
}

// insertStockFundPurchase godoc
// @Summary Insert some stocks fund purchase
// @Description insert new stocks fund purchase
// @Accept json
// @Produce json
// @Router /stocks-funds/purchases [post]
func (s *server) insertStockFundPurchase(c echo.Context) error {
	log.Debugf("[API] Inserting stock fund purchase")

	data := wallet.NewStockFund()

	if err := c.Bind(data); err != nil {
		log.Errorf("[API] Error on bind: %v", err)
		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
	}

	if err := c.Validate(data); err != nil {
		log.Errorf("[API] Error on validate: %v", err)
		return echo.NewHTTPError(http.StatusUnprocessableEntity, err.Error())
	}

	result, err := s.db.InsertStockFundPurchase(data)
	if err != nil {
		log.Errorf("[API] Error on insert: %v", err)
		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
	}

	return c.JSON(http.StatusOK, result)
}

// updateStockFundPurchaseByID godoc
// @Summary Update some stocks fund purchase
// @Description update new stocks fund purchase
// @Accept json
// @Produce json
// @Router /stocks-funds/purchases/{id} [put]
// @Param id path string true "Purchase id"
func (s *server) updateStockFundPurchaseByID(c echo.Context) error {
	id := c.Param("id")
	log.Debugf("[API] Updating stock fund purchase with id %s", id)

	data := wallet.NewStockFund()

	if err := c.Bind(data); err != nil {
		log.Errorf("[API] Error on bind: %v", err)
		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
	}

	if err := c.Validate(data); err != nil {
		log.Errorf("[API] Error on validate: %v", err)
		return echo.NewHTTPError(http.StatusUnprocessableEntity, err.Error())
	}

	result, err := s.db.UpdateStockFundPurchaseByID(id, data)
	if err != nil {
		log.Errorf("[API] Error on update: %v", err)
		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
	}

	if result.MatchedCount != 0 {
		return c.JSON(http.StatusOK, result)
	}

	return c.JSON(http.StatusNotFound, "Stock fund purchase not found")
}