# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  wallet.FII:
    properties:
      brokerId:
        type: string
      commission:
        type: number
      date:
        type: string
      id:
        type: string
      itemType:
        type: string
      portfolioId:
        type: string
      price:
        type: number
      shares:
        type: number
      symbol:
        type: string
    required:
    - brokerId
    - date
    - itemType
    - portfolioId
    - price
    - shares
    - symbol
    type: object
  wallet.Portfolio:
    properties:
      costBasics:
        type: number
      gain:
        type: number
      id:
        type: string
      items:
        additionalProperties:
          $ref: '#/definitions/wallet.PortfolioItem'
        type: object
      name:
        type: string
      overallReturn:
        type: number
    required:
    - id
    - name
    type: object
  wallet.PortfolioItem:
    properties:
      averagePrice:
        type: number
      brokerId:
        type: string
      change:
        type: number
      closingPrice:
        type: number
      commission:
        type: number
      costBasics:
        type: number
      gain:
        type: number
      itemType:
        type: string
      lastPrice:
        type: number
      lastYearHigh:
        type: number
      lastYearLow:
        type: number
      name:
        type: string
      purchases:
        $ref: '#/definitions/wallet.PurchasesList'
        type: object
      sales:
        $ref: '#/definitions/wallet.SalesList'
        type: object
      sector:
        type: string
      segment:
        type: string
      shares:
        type: number
      subSector:
        type: string
    type: object
  wallet.PurchasesList:
    items:
      type: object
    type: array
  wallet.SalesList:
    items:
      type: object
    type: array
...
  wallet.StockFund:
    properties:
      brokerId:
        type: string
      commission:
        type: number
      date:
        type: string
      id:
        type: string
      itemType:
        type: string
      portfolioId:
        type: string
      price:
        type: number
      shares:
        type: number
      symbol:
        type: string
    required:
    - brokerId
    - date
    - itemType
    - portfolioId
    - price
    - shares
    - symbol
    type: object
...
host: localhost:8889
info:
  contact:
    name: API Support
    url: https://github.com/mfinancecombr/finance-wallet-api
  description: mfinance Wallet API data.
  license:
    name: BSD 3-Clause
    url: https://opensource.org/licenses/BSD-3-Clause
  title: MFinance Wallet API
  version: 0.1.0
paths:
...
  /fiis/purchases/{id}:
    get:
      consumes:
      - application/json
      description: get FII purchase data
      parameters:
      - description: Purchase id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.FII'
      summary: Get FII purchase by ID
    put:
      consumes:
      - application/json
      description: update new FII purchase
      parameters:
      - description: Purchase id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      summary: Update some FII purchase
...
  /fiis/sales/{id}:
    get:
      consumes:
      - application/json
      description: get FII sale data
      parameters:
      - description: Sale id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.FII'
      summary: Get FII sale by ID
    put:
      consumes:
      - application/json
      description: update new FII sale
      parameters:
      - description: Sale id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      summary: Update some FII sale
  /portfolios:
    get:
      consumes:
      - application/json
      description: get all portfolio data
      parameters:
      - description: filter by year
        in: query
        name: year
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/wallet.Portfolio'
            type: array
      summary: List all portfolios
  /portfolios/{id}:
    delete:
      consumes:
      - application/json
      description: delete some portfolio by id
      parameters:
      - description: Portfolio id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      summary: Delete portfolio by ID
    get:
      consumes:
      - application/json
      description: get all portfolio data
      parameters:
      - description: Broker id
        in: path
        name: id
        required: true
        type: string
      - description: filter by year
        in: query
        name: year
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.Portfolio'
      summary: Get a portfolio
    put:
      consumes:
      - application/json
      description: Update some portfolio by id
      parameters:
      - description: Portfolio id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      summary: Update portfolio data by ID
...
  /purchases/{id}:
    delete:
      consumes:
      - application/json
      description: delete some purchase by id
      parameters:
      - description: Sale id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      summary: Delete purchase by ID
...
  /sales/{id}:
    delete:
      consumes:
      - application/json
      description: delete some sale by id
      parameters:
      - description: Sale id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      summary: Delete sale by ID
...
  /stocks-funds/purchases/{id}:
    get:
      consumes:
      - application/json
      description: get stocks fund purchase data
      parameters:
      - description: Purchase id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.StockFund'
      summary: Get stocks fund purchase by ID
    put:
      consumes:
      - application/json
      description: update new stocks fund purchase
      parameters:
      - description: Purchase id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      summary: Update some stocks fund purchase
...
swagger: "2.0"